extract jwt from header
